3 features added in self service portal
Requestors should be able to search tickets and sort.

Tickets on the user side sort from oldest to newest.  Some of clients have 30 open tickets and it would make sense to list the newest tickets first to save them the scroll

Add the EXPORT button to the Client side.  Some of the clients like to export tickets for their own offline review.  Right now that needs to be done by our agents.

Signed-off-by: Navaneeth <navaneeth@freshdesk.com>