JS based Enhancements for Ticket List view
Bringing Assign to None and Myself at the top of list.
The Agent list will not be loaded via AJAX.
Adding search when the list of options exceeds 10.
Cache Busting for Profile Image used in the imgerror function.
Simplified code to handle Agent/Status/Priority Assigning in the same handler.
Fixed a bug in TicketList where it was not restoring tickets, when clicking on ticket 2nd time and returning back.

Signed-off-by: Thanashyam Raj <thanashyam@gmail.com>