Tracking duplicate Like/Dislike and Hits in Solution Articles
Tracking duplicate like/dislike :

Voting container is hidden for users who have already voted.

For a non logged-in user, the thumbs up/ thumbs down is incremented when he votes for the first time.

For a logged-in user, if the user is an agent, thumbs up/thumbs down is not incremented when he votes.

For a logged-in user, thumbs up/thumbs down is incremented when he votes for the first time. Also, a record is created in the votes table for the same.

For a logged-in user, if he attempts to vote a second time, the votes table is checked for his choice during his first vote. If there is a change in choice, the record in votes table is updated and thumbs up/thumbs down is incremented and thumbs down/thumbs up is decremented. If there is no change, nothing is done.

Tracking views :

hit column has been added to the solution_articles table to track the number of times an article has been viewed.

Signed-off-by: sravii <sravya@freshdesk.com>