[#4793], [#4932] Displayed appropriate error message while moving folders between categories
When a folder is moved between categories, and the target category already has a folder with that name, the error message being displayed has been changed.

The uniqueness validation for name of category, forum category, folder and forum made case insensitive.

Eg:
Previously, If My Folder exists and we try to create My folder, it passes the Rails validation and a table insert is attempted.
And the MySQL validation fails resulting in "We're sorry, but something went wrong." error page being shown.
Now, the validation will fail at the Rails level and the appropriate error message is displayed.

Also fixed a few translation issues for Dutch.

Signed-off-by: sravii <sravya@freshdesk.com>