Fixed issue in solution folder creation
Previously, when we choose a different category while creating a new folder, the folder gets added to the wrong category.

Now, the new folder will be added to the chosen category if any category is chosen.
Also, the redirection path has been changed to redirect to the category to which a new folder has been added.

Signed-off-by: sravii <sravya@freshdesk.com>