[#5043] Enabling canned responses' visibility to multiple groups - Back-End changes
Canned responses can be made visible to multiple groups. There by agents of the respective groups can use canned response.

In this phase, only back end changes are done. Hence canned response's behaviour remains the same as of now. The new tables are updated along with the previous tables in the database. And also the data for the view is still being fetched from old tables.

Signed-off-by: vinodhkumarS <vinodhkumar@freshdesk.com>