[#4755] Fix the updated_at in Solution::Articles
The "updated_at" timestamp is being updated whenever the article is being re-ordered or has other changes.

But the customers want the timestamp when the article's contents were edited, not updating timestamps for other changes.

A new column has been added - "modified_at" which will hold the timestamps of the edited article. This new column will be updated only when title or description of an article changes. This timestamp won't be updated if the article's folder or tags or status or any other properties are changed.

Signed-off-by: Sony Mathew <sony@freshdesk.com>