[#5232] [#177012] Security Issues : Support discussions topic posts edit
The vulnerability was that anyone with the correct url could edit any others posts in the support forums, provided he maintains a session as logged in customer. The URL is
https://support.freshdesk.comsupport/discussions/topics/TOPIC_ID/posts/POST_ID/edit?_method=get
where you have to give the Topic ID and Post ID manually. This will render a partial with a text area with post content and you could edit the content and update it.

Signed-off-by: Sony Mathew <sony@freshdesk.com>