[#5162] Changed the message that appears when a new topic or reply is posted
When moderation is set to : Moderate only suspicious topics and replies or Moderate all topics and replies,

For new topic : Thank you for starting a new topic. Your topic will appear in the forum, after a standard moderation process.

For new reply : Thank you for posting. Your reply will appear in the topic, after a standard moderation process.

When the moderation is set to : Do not Moderate

For new topic : Thank you for starting a new topic. Your topic will appear in the forum, after checking for spam.

For new reply : Thank you for posting. Your reply will appear in the topic, after checking for spam.

Signed-off-by: sravii <sravya@freshdesk.com>