[#5187] Issue in tags associated with Solutions
Previously, when an API call was made for creating/updating a solution and the params for tags are wrongly specified (improper structure), an error was thrown.

Now, when the params for tags are wrongly specified, it doesn't throw an error. Instead, the solution is created without any changes to the tags.

Signed-off-by: sravii <sravya@freshdesk.com>