[#7369]Canned Response for multiple groups[Phase2]
In this phase, only back end changes are done. Hence canned response's behaviour remains the same as of now.
The difference is the data is being fetched from the new tables rather than old access tables for canned response.

Note : First phase of this feature was pushed in with the bug-id #5043.

Signed-off-by: vinodhkumarS <vinodhkumar@freshdesk.com>