[#7682] Custom ticket views for multiple groups[Phase 1]
Custom ticket views can be made visible to multiple groups. There by agents of the respective groups can use the custom ticket view.

In this phase, only back end changes are done. Hence ticket views' behaviour remains the same as of now. The new tables are updated along with the previous tables in the database. And also the data for the view is still being fetched from old tables.

Signed-off-by: vinodhkumarS <vinodhkumar@freshdesk.com>