[#7491] Delete Confirm Dialogue box in Solutions & Forums
Previously in the edit pages of solution category, solution folder, solution article, forum category, forum and topic, when you click on the delete button, it will show a dialogue box asking for confirmation.
Now a text field has been added in this pop-up(except in case of topic delete), so that the delete button will become active only if the user enter the first five characters of the title of the item he/she is going to delete.

commit 04a2514eb4de9ec7c059efcd6f47bf1ab4b19c9d
Author: Sony Mathew <sony@freshdesk.com>
Date:   Mon Jan 5 16:20:03 2015 +0530

    [#7491] Delete Confirmation Message Change

    Changed the displayed message in the delete confirmation popup from "This action is undoable"  to "This action is irreversible".

    Signed-off-by: Sony Mathew <sony@freshdesk.com>

commit 61477058b0f7d0dd60968a7eec498d453c82a81b
Author: Sony Mathew <sony@freshdesk.com>
Date:   Wed Dec 31 18:06:39 2014 +0530

    [7491] Cleaning up  delete_confirm.js

    Signed-off-by: Sony Mathew <sony@freshdesk.com>

commit 3d2f19846fe8157d11fad9ff3fe7b0c776e36211
Author: Sony Mathew <sony@freshdesk.com>
Date:   Wed Dec 31 17:18:27 2014 +0530

    [#7491] Delete Confirm Dialogue box in Solutions & Forums

    A confirmation popup asking to type in the name of the item to be deleted in Solutions & Forums and UI revamp has been done.

    Signed-off-by: Sony Mathew <sony@freshdesk.com>

Signed-off-by: Sony Mathew <sony@freshdesk.com>