[#8194] Remove Topic/Reply editing in forums from the support portal
Spammers use this way to inject malicious content in the forums.
They create the topic or a reply with simple content, which is most likely to get published. Once they are published, they edit them and update them with spammy content.

To avoid this scenario and help our customers maintain clean, spam-free forums, we are going to do away with edit option in topics and posts.
The agent will still have the edit option in the helpdesk view.

Signed-off-by: Thanashyam Raj <thanashyam@gmail.com>