[#8907] Kbase Forwarded Articles not appearing in Portal
When a reply is forwarded from a Ticket to Kbase from our app, it adds freshdesk_quote class to the contents to make it look like a blockquote.
This class is retained even when solution is created. When this solution is viewed in the agent side, there is no style for this class, and so everything looks good.
But in the portal, this class has the style: display: none;. This hides the entire content in the portal.

After the fix, freshdesk_quote in solution articles (portal side) will not be hidden.

Signed-off-by: Subin <subin@freshdesk.com>