[#8924] New Topic Captcha mistake & Topic content
In the portal new topic page, when you enter the wrong captcha and submit, you are sent back to the same new topic page but without the content loaded (title will be there but not the body).

Only when signed in as an end user and not as an agent.

Signed-off-by: Subin <subin@freshdesk.com>