Log tagging with request ID uuid.
[#9337]
Rails 3 brings a way to tag all the requests with a uuid. The same uuid is then returned as X-Request-ID header response and also written to the logs. With this the logs can be easily filtered out and helps in faster debugging of the issues. This is especially useful for API and support teams.