Facebook Rake - bug fix
The pages are fetched properly for the first shard, but when we query the second shard, Account.current is set to an account in the previous shard so no pages are fetched.

Remove account.make_current instead of doing Account.reset_current_account as every call is scoped by account.