Language translation tabs in article show/new page
Scenario 1 : When an article is being created
We show the primary language tab and a dummy dropdown which when hovered upon, shows the message - “Create article in your primary language to select other language translations”

------------------------------------------------------------------------------------------------------------------------------------------------

WHEN AGENT’S PROFILE LANGUAGE IS DIFFERENT FROM ACCOUNT PRIMARY LANGUAGE

Scenario 2 : When article exists but no translations exist
We show the primary language tab, and then the tab for agent’s profile language. After this, we show a dropdown with count and list of remaining languages.

Scenario 3 : When article exists and only one translation exists, that too in agent’s profile language
We show the primary language tab, and then the tab for agent’s profile language. After this, we show a dropdown with count and list of remaining languages.

Scenario 4 : When article exists, < 4 (tab limit - 1) translations exist and agent’s profile language doesn’t have a translation
We show the primary language tab, then the tabs for languages in which translations exist and then we show the tab for agent’s profile language. After this, we show a dropdown with count and list of remaining languages.

Scenario 5 : When article exists, < 4 (tab limit - 1) translations exist and agent’s profile language has a translation
 We show the primary language tab, then the tab for agent profile language and then tabs for other languages in which translations exist. After this, we show a dropdown with count and list of remaining languages.

Scenario 6 : When article exists, > 4 (tab limit - 1) translations exist and agent’s profile language doesn’t have a translation
We show the primary language tab, then the tabs for languages in which translations exist until tab limit is reached. After this, we show a dropdown with count and list of remaining languages.

Scenario 7 : When article exists, > 4 (tab limit - 1) translations exist and agent’s profile language has a translation
We show the primary language tab, then the tab for agent’s profile language and then tabs for other languages in which translations exist until tab limit is reached. After this, we show a dropdown with count and list of remaining languages.

Scenario 8 : When article exists, < 3 (tab limit - 2) translations exist, agent’s profile language doesn’t have a translation and a particular language is chosen from dropdown
We show the primary language tab, then tabs for other languages in which translations exist, then we show the tab for the particular language chosen and then the tab for agent’s profile language. After this, we show a dropdown with count and list of remaining languages.

Scenario 10 : When article exists, > 3 (tab limit - 2) translations exist, agent’s profile language doesn’t have a translation and a particular language is chosen from dropdown
We show the primary language tab, then tabs for other languages in which translations exist and then we show the tab for the particular language chosen. After this, we show a dropdown with count and list of remaining languages.

Scenario 11 : When article exists, agent’s profile language has a translation and a particular language is chosen from dropdown
We show the primary language tab, then the tab for agent profile language, then tabs for other languages in which translations exist and then we show the tab for the particular language chosen. After this, we show a dropdown with count and list of remaining languages.

---------------------------------------------------------------------------------------------------------------------------------------------------------

WHEN AGENT’S PROFILE LANGUAGE IS SAME AS ACCOUNT PRIMARY LANGUAGE

Scenario 12 : When article exists but no translations exist
We show the primary language tab. After this, we show a dropdown with count and list of remaining languages..

Scenario 13 : When article exists, 1 or more translations exist
We show the primary language tab, then the tabs for languages in which translations exist until tab limit is reached. After this, we show a dropdown with count and list of remaining languages.

Scenario 14 : When article exists and a particular language is chosen from dropdown
We show the primary language tab, then 3 (tabs limit - 2) tabs for languages in which translations exist and then the tab for that particular language. After this, we show a dropdown with count and list of remaining languages.

---------------------------------------------------------------------------------------------------------------------------------------------------------

Scenario 15 : When article exists, number of languages in the dropdown is only 1
We show the primary language tab, then the tabs for languages in which translations exist until tab limit is reached. After this, we show the tab for the remaining language.

NOTE : Languages are shown in the order in which they are present in the account’s supported languages.

Signed-off-by: Sravya <sravya@freshdesk.com>