Assumable agent tocheckin
Assumable agents list was loaded on page load. This loading has now been done with an ajax call made on the click of the user avatar.

When there is a delay in the loading "Loading agents..." message will be seen

Javascript portions(corresponding to switch agent and other js in the page) have been moved to a separate js file

Availability for round robin toggle, toggle shortcuts and viewing shortcuts are the other impact sread to be tested

Signed-off-by: iswaryar <iswarya.r@freshdesk.com>