[#16137] Sidekiq - Redis Connection timout fix
Redis connection timeout for select_all job causing issues in reports/ES:

We are getting redis connection timeout error in sidekiq instances listening to select all queues during the push of sidekiq jobs for reports/ ES in ticket callback. This is happening as Sidekiq batches extensively uses redis and there seems to be no available redis connection in connection pool for sidekiq client for pushing new jobs. The solution to this is to increase the connection pool size of redis.

Signed-off-by: manojcg <manojcg@freshdesk.com>