Bug 16145 - Stop mailbox job execution after 5 minutes to avoid process holding
Set maximum job run time for mailbox queue jobs to 5 minutes.

Currently there is no auto timeout configured in our delayed job process (single thread) so when there is an issue in smtp delivery operation it is running forever.

This request is to apply timeout in delayed job process so that it won't block other rows in queue.

Signed-off-by: Athi Ruban <athi@freshdesk.com>