[# 16512] - Mobile Login controller Fix
This fix will allow mobile login for multiple pods.

This commit also has one other inclusion of before filter which is not related to mobile, but it is related to shard selection for taking SSO from freshopsadmin, for which no testing is required and it will not have any other impact.