Fix for bug 16670
Before fix for bug 16457,

Issue: customer A and customer B belong to company C. customer A is a client manager. A has chosen to see tickets of B in the ticket list view and is viewing the same. Parallely when the admin removes the company for customer A, and if this ticket list view is refreshed(in customer login), A is still able to see B's tickets.

This was fixed. But the scenario where A's company is changed wasn't handled in the fix. Fixing it here.

Signed-off-by: iswaryar <iswarya.r@freshdesk.com>