[#16646] Handling multiple email addresses in the "To"
An agent forwards a ticket conversation to a third party. Now, the third party responds by including one more email id in the "To" address] and a few email ids in the CCs. This gets added as a private note on the ticket.

Now, when the agent replies to the third party ( to continue the discussion), the second "To" address gets dropped. CCs remain intact.

This has been fixed.

Signed-off-by: Sravya <sravya@freshdesk.com>