Fix the bug in clearing reply cc
Issue:
Reply to an email triggered by automation(Email to requester/group/agent) replaces the existing reply cc of the ticket.
Reply to the reply to an email notification or automation also replaces the existing reply cc

Solution:
Reply to an email triggered by automation/notification should be appended to existing reply cc and not replaced.
Reply to the reply to an email notification or automation should also behave the same way.

Signed-off-by: Divya <divya@freshdesk.com>