Fix for "Mail not sent to ticket watchers, if ticket is resolved/closed by system rules"
USE CASE: notify_wathcers method notifies the watchers when the current user is not the subscribed user. To accomplish this, it checks for subscription.user.id != user.current.id.

ISSUE: But when an automation closes/resolves a ticket, user.current is nil and user.current.id raises an error and thus subsequent callbacks are not invoked and also the mail is not sent.

FIX: user.current nil scenario is handled. If user.current is nil, the mail is sent as "An automation rule" closed your ticket, else the user name is used.
Signed-off-by: iswaryar <iswarya.r@freshdesk.com>