Bug [#17522]
Summary: On updating "email" and "other_emails" with existing email id, the filed in the error message shows only "other_emails" and the message in the error includes both primary and secondary email. Expected behaviour is the that API has to return 2 errors one on email and another on other_emails along with the corresponding emails.

Reviewers: soorya, rashmi

Reviewed By: soorya, rashmi

Differential Revision: http://wall.freshdesk.com/D495