Fix for  [#18072]: When an agent tries to merge three contacts with two contacts with same work phone and mobile number and one contact with same work phone an error message is getting displayed
The counter for each field (twitter, mobile, work phone) is incremented when the primary and secondary contact have different values. It is not incremented when they are the same in which case we allow the merge ot happen. But when one of the contacts doesn't have the field, "" and the value of the other is compared and it is considered unequal. So the counter is incremented and thus validation error shown. Have handled the empty case now

Signed-off-by: iswaryar <iswarya.r@freshdesk.com>