[#19265] Optimize i18n usage
Currently all our localization strings (~250K objects) live in the
Ruby heap. Since we do know that these strings are never going to be
garbage collected, we can move them off of Ruby heap and then manage
them ourselves.

This change is using i18nema to do that.

This will translate to the following benefit

 - About 10% reduction in GC (full)
 - Less CPU usage
 - Fast app loading/bootup
 - Marginal speed improvement in locale string lookups