[#18514] Merge tickets - Add recipients checkbox
The "Add recipients" checkbox should be checked by default when all the tickets are from the same requester. Everytime a ticket is added to or removed from the merge tickets list, this check has to happen again. But if the customer checks or unchecks the checkbox, we won't change its state for any subsequent tickets added to or removed from the merge list.

Signed-off-by: Sravya <sravya@freshdesk.com>