[#20408] System activities - Increasing the any system rule name size limit to 100 in activities
System activities - Increasing the any system rule name size limit to 100 in activities.
This includes the length of the tooltip string present in system activity for Dispatch'r rule, observer and supervisor rule.
Note: Tooltip length for will be 100 only if the rule is deleted, otherwise it will show the actual length the rule name.

Signed-off-by: Ramkumar S <sram2020@gmail.com>