[#20659] Revert agent collision url for mobile
Android mobile clients have issues connecting to new node servers since the new servers use session based cookies and the client doesn't have a way to persist the cookie on the requests. This requires code change on mobile. In the mean time, the mobile clients will be connecting to old endpoint till this gets fixed.

Signed-off-by: Manoj C G <manojcg@freshdesk.com>