[#20365] Bug fix Bulk Reply
If an agent replies to a ticket from the list page, using bulk actions, the reply is sent only to the requester and not to the contacts in the CC.

After this, if the agent opens the ticket and replies to in from the details page, since the cc contact is removed in the previous reply, the cc doesn't appear in the current reply as well.

expected behaviour:
send bulk reply form list view to requester and people in reply cc.

Signed-off-by: Manu <manu@freshdesk.com>