#20607 Replace invalid utf-8 characters while creating new mail
We replace the invalid UTF-8 characters with null string when they exist in "subject" and "headers" part of the email. But the invalid characters are also present in other parts of email such as "text" and "html". And this raise NewRelic exception often. It can be avoided by following the same policy to all parts of the email.