[22182] - Mail header information duplication
Description of the Bug:

information in mail header added multiple times if the emails are generated inside a loop operation( ex, if emails for SLA notification are generated for multiple agents in a loop). Inside the loop, the same mail object is being used every time and the new header is being appended to the existing one.

Description of the Fix:

Email headers are being reset every-time inside the loop

2. Adding Quotedtext detection fix [919667]

3. [859243] Forward email square bracket Issue

4. Adding the from_email to custom_variables of email headers