Bug Id : 20389
ISSUE: When the helpdesk name is changed, it is not getting updated in an email signature sent for ticket export

REASON: account.helpdesk_name, account.name and portal.name have been redundant. The issue happens because when we update helpdesk name in admin>helpdesk page, the portal’s name gets changed but the name and helpdesk_name fields in account table are not altered. So the change is not reflected in account>helpdesk page nor does it reflect in some email signature

SOLUTION: going forward, we will provision for having helpdesk_name and portal_name to be different.That is, we can store unique value for each of the fields.The reason is because we have customers who will have their helpdesk name as their umbrella brand name and the portal name will be their specific product names.

FIX: so for now, when the name is changed in account>helpdesk page, account.name and account.helpdesk_name is changed. Portal name is not updated.

CONCERN: since so far the three fields account.name, account.helpdesk_name and portal.name (or account.portal_name) have been used synonymously/redundantly, some places use portal name and some account name. We are yet to decide which page should display which value. We are not doing any changes now. It is has been decided with the product team that it will be done when there is a need
Signed-off-by: iswaryar <iswarya.r@freshdesk.com>