[JIRA] (FDCORE-708) - Social DM Issue - Customer Responded tag not being shown
If the recent requester_responded_at is more recent than the recent agent_responded_at, "customer_responded" label comes up against the particular ticket in ticket_list view.

In case of customer responses through direct messages from fb and twitter, the "customer_responded" labels didn't show up.

Reason: The customer responses through DMs are saved as private notes. and the recent requester_responded_at is not updated in case of private_notes.

Fix: altered the code to update recent requester_responded_at - it won't be updated only when it is a private note that is added by the agent.. the incoming private notes  that are through social DMs will be considered for updating the field
Signed-off-by: iswaryar <iswarya.r@freshdesk.com>