Email  failure notification
When an email fails to deliver, the same needs to be shown as a notification in the ticket conversation view.

The following is the scope of this:
Ticket details page - Description message (in case of outbound emails / Phone tickets)
Ticket details page - Agent replies
Ticket details page - Agent forwards
For all of the above, we will show failures in case of To and CC addresses. BCC addresses are not included in this scope.