FDCORE-6077 -  Add requester/contact field in Edit ticket
This commit contains the API change for adding the contact field in the Edit Ticket page, where the requester of the ticket can be changed. Also, If the account has one contact multiple companies feature enabled, the company field will also be displayed if the new contact has multiple companies associated with it.