FD-4173 Notify Freshid failures/Exception via SNS (#1418)
JIRA Id(s) : FD-4173

Description :
< Please describe the changes so that the reviewer has enough details >
It is difficult to keep track of failures occurred when interacting with Freshid. Freshid APIs are being used in lot of places(controllers). We couldn't search for errors in newrelic. If the error is notified in email, we can proactively check the affected account and resolve the issue.

Why is this change necessary?

How does it address the issue?
Notifying via SNS