FD-364-[summary-app] (#1281)
FD-364-[summary-app]
Summary app is not available in old UI. Summaries created in the new UI will not appear in the old UI.

To add a special note - which appears right below the ticket description, before all notes and replies, and even before the expand button. This note would act as the Summary where agents can dump all they know about an issue.