[FD-3810] Fix for UI breakage in FreshID enabled accounts (#1522)
JIRA Id(s) : FD-3810

Description :
Why is this change necessary?
For Freshid enabled accounts there is small gap coming below the every tab.
How does it address the issue?
This was due to the 1x1 image that was added to keep FreshID session alive when there is activity in Freshdesk. Hence, display: none is added to the image tag to fix this.