FD-4529 FD-4315  FD-4631 Change the format of automation log (#1458)
Adding logs for tracking performance of supervisor
Change the automation log format
Add time taken to execute a dispatcher, observer and supervisor rule

Description:
Change the automation log format after discussing with Baikal team. Baikal team was suggesting a standard format which will help them to easily parse the data and load it in their DB

Add time taken to execute a dispatcher, observer and supervisor rule in automation.log, so that we can later send the log to baikal and run query to find out rules which are affecting the performance.