FD-4587 Apigee integration (#1589)
JIRA Id(s) :
https://jira.freshworks.com/browse/FD-1037

Description :
Changes for apigee integration. If api throttling has been done in apigee then we don't want to throttle again in fd_api_throttler. If it is the new instance "apigee-api" then fd_api_throttler middleware will not be included. So no issues.
But in case of premium_api_app instance we have to handle this situation. Add launch party feature :apigee and a check in fd_api_throttler.

Why is this change necessary?
Because api throttling has already been done in apigee. It is like doing the same job again.

How does it address the issue?

Any side effects or dependencies?
In case if the feature is enabled for account but apigee is down or account details not present in apigee then throttling will not happen. In this case we have to change node[:apigee][:enabled] to false in stack settings and rollback this launch party feature.