[FD-5695] Bug fix - Onboarding task complete in FreshID new account signup (#1605)
JIRA Id(s) : FD-5695

Description :
Why is this change necessary?
When existing FreshID user signs up for new account, 'Invite your Team' and 'Add an automation rule'(Falcon) are shown as complete by default.
How does it address the issue?
This was due to explicitly activating the account if user is active in FreshID. Due to which the update observer ran on account and thus create observer did not consider the account to be a new one. The fix for this is to change the condition to check for new account signup.
Any side effects or dependencies?
Need to check normal and FreshID signups. Also, need to check the completion of onboarding tasks process.