[FD-5569] FreshID bug fix for convert to agent (#1599)
JIRA Id(s) : FD-5569

Description :

Why is this change necessary?
If a new agent is added in FreshID enabled account, activated, converted to contact and then converted again to agent, it doesn't happen. Only if password for contact is set, we could convert to agent.
How does it address the issue?
Reset the user active status to false if there are no password/authorizations while converting a contact to agent in FreshID enabled account.