[FD-1780] Saved canned responses with placeholders fix (#1620)
The customer is adding the placeholder ticket.id in the link which is inserted in the image or the text in the canned response. He is saving the sample link "ww.website.com?ID={{ticket.id}}" in the canned response. However, after saving and going back to the insert link content it is available as "%7B%7Bticket.id%7D%7D"

The customer has used the placeholder on his website before, however, he is having the troubles now as the URL is getting changed.