FD-6778 Misconfigured dkim records handled
Due to sendgrid's recent API limitation, some accounts DKIM records are partially created/misconfigured.
So if the DKIM configure fails, we tried to retrieve the partial record which already got created and configure it on freshdesk side.
On deletion, sometimes restoring dkim configure option for domain not happening. It was handled.