[FD-4837, FD-7782] Toggle FreshID when SSO is enabled/disabled (#1607)
JIRA Id(s) : FD-4837, FD-7782

Description :
Why is this change necessary?
If SSO is enabled, we need to rollback FreshID feature and delete the account and user mapping. Until we finish that, we shouldn't allow customer to disable SSO. New default agent password policy must be created, password reset emails need to be sent from Freshdesk and we need to remove the agent invitation template.

When he disables SSO, we need to enable FreshID for the account. Until the syncing of data with FreshID completes, we shouldn't allow him to turn on SSO. Agent password policies must be deleted and password reset emails will no longer be sent from Freshdesk. Also, we need to add a new email notification - Agent invitation.

How does it address the issue?
Whenever, SSO is enabled or disabled we revert/migrate users to FreshID. SSO toggle option is temporarily disabled when the migration process is in progress.