[FD-6962] Fixes redirection urls for preview in portal customisation (#1648)
For solution category home template, we are redirecting to solution home page, we should redirect to solution category show page.
For discussions category home template, we are redirecting to discussions home page, we should redirect to discussion category show page.
For article list template and article view template, the scopes being used to generate the redirection url are throwing an error, which goes to exception handling and redirects to folder show and article show page respectively with id as the template id. These templates should instead redirect to a public folder show page and published article show page.