[FD-7495] Skip FreshID user creation for custserv@freshdesk.com (#1740)
JIRA Id(s) : FD-7495

Description :

Why is this change necessary?
custserv@freshdesk.com is not in use. So, we need add a check in Freshdesk to not to create this user in FreshID to avoid timeouts.
How does it address the issue?
Do not create user in FreshID for custserv email.