FD-5409 validate redirect_to url in announcements (#1941)
JIRA Id(s) :
FD-5409

Description :
Our application is vulnerable to URL redirection attack as it does not validate the page request in the ‘redirect_to’ parameter to check if the page belongs to the Freshdesk domain

Why is this change necessary?

How does it address the issue?
If a domain other than freshdes domain / cname is given, it won't be redirected to that page. Instead it will be render 404 page. In logger the error will be mentioned.