Fd 10490 helpkit (#2129)
JIRA Id(s) :
FD-10490

Description :
Why is this change necessary?
Twitter has enforced whitelisted_url in authorisation. This is to handle that.

How does it address the issue?
Hereafter, authorisation will be made using integrations_url which is login.freshdesk.com and it is whitelisted in our twitter app.

Any side effects or dependencies?
Need to check twitter login.