[FD-7515] Prevents recording survey response from email scanner (#2204)
We'll be having a launch party feature for accounts which have issues due to email scanner submitting survey responses.

Current flow : When an option is chosen in the email, another page is rendered and survey response is recorded via an Ajax call during page load. In the page that is rendered, on clicking submit additional feedback can be recorded.

For accounts with the launchparty enabled, we'll prevent the response from being recorded and instead record the response on clicking Submit on the page that is rendered.