FD-3667 - [Mint] Inconsistent CC auto removal (#2324)
cc emails in the Reply template contains the entire list of cc emails form the ticket. Made changes to the reply template, so that it returns the reply_cc instead of current_cc. The problem was due to flipped if condition in the function returning the reply_cc_emails in contrast to the old UI.