[FD-14727] Saml encrypted assertion (#2769)
This is being done to support encrypted assertions in SAML authentication.
It was decided to have a single common key for encrypted assertions.

What does this code do

reads the private key certificate and sets it as the sp's private key in SAML settings.
There is no need for us to check if encrypted assertions are enabled or not as the gem is able to get this information from the Response. If the response has an encrypted assertion then the private key will be used to decrypt the assertion.

moved the settings into SamlResponse initialize instead of setting it later as the response gets decrypted during initialization.