[Bug] [FD-62209] Introducing Shoryuken Middleware for Resetting Thread variables (#9275)
As we've upgraded Shoryuken, one hidden thing is that it is using concurrent_ruby for multi-threaded processing.
Which means we should take care of Thread sanity.

Twitter ticket creation involves setting a User as current. And the thread variable is not getting reset.
When FB ticket is created via Shoryuken, as it eventually gets enqueued for Dispatchr, the current user is sent in the parameter of the Sidekiq job. Since it is a different job, the user might belong to a different account.

When sidekiq processes, it tries to load this user which will not be present in the current account. This causes dispatcher to halt there.

The reason such a scenario might not be found in a lot of Shoryuken workers is that, very rarely we set thread variables apart from Account, from within a Shoryuken job.

Fix: Introduce a middleware in Shoryuken that will reset all the thread variables