[Technical Story] [FD-64535] Switch to mysql2 adapter for tracing (#9564)
Remove Active Record and Active Support adapters and switch to mysql2 adapter as mysql2 adapter had lesser over head compared with the AR adapter.

Also updating the ejson files for selective enablement in production by introducing env variable.

We will sample the request only when traceparent header is set and sent from Edge. On the SDK side we are not doing any client sampling so setting probability to 0.